Rename uvm-install to uvm-install2 and revert to original command
