+++ install/uvm-install2/Cargo.toml
@@ -1,10 +1,11 @@
 [package]
-name = "uvm-install"
+name = "uvm-install2"
 version = "2.0.1"
 authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
 
 [dependencies]
 flexi_logger = "0.9.2"
+indicatif = "0.9.0"
 log = "0.4.5"
 console = "0.6.1"
 serde = "1.0"
