+++ install/uvm-install2/src/main.rs
@@ -1,20 +1,22 @@
 #[macro_use]
 extern crate uvm_cli;
-extern crate uvm_install;
+extern crate uvm_install2;
 extern crate flexi_logger;
 extern crate console;
+
+#[macro_use]
+extern crate log;
 
 use console::style;
 use std::process;
 use console::Term;
-use std::io::Write;
 
 const USAGE: &'static str = "
-uvm-install - Install specified unity version.
+uvm-install2 - Install specified unity version.
 
 Usage:
-  uvm-install [options] <version>
-  uvm-install (-h | --help)
+  uvm-install2 [options] <version>
+  uvm-install2 (-h | --help)
 
 Options:
   -a, --all         install all support packages
@@ -26,20 +28,21 @@ Options:
   --windows         install windows support for editor
   --desktop         install desktop support (linux, windows)
   -v, --verbose     print more output
+  -d, --debug       print debug output
   --color WHEN      Coloring: auto, always, never [default: auto]
   -h, --help        show this help message and exit
 ";
 
 fn main() -> std::io::Result<()> {
-    let mut stdout = Term::stderr();
-    let options:uvm_install::Options = uvm_cli::get_options(USAGE)?;
-    uvm_install::UvmCommand::new().exec(options)
+    let stdout = Term::stderr();
+    let options:uvm_install2::Options = uvm_cli::get_options(USAGE)?;
+    uvm_install2::UvmCommand::new().exec(options)
         .unwrap_or_else(|err| {
-            let message = format!("Unable to install");
-            write!(stdout, "{}\n", style(message).red()).ok();
-            write!(stdout, "{}\n", style(err).red()).ok();
+            let message = format!("Failure during installation");
+            stdout.write_line(&format!("{}",style(message).red())).ok();
+            info!("{}", &format!("{}",style(err).red()));
             process::exit(1);
         });
 
-    stdout.write_line("Finish")
+    stdout.write_line(&format!("{}", style("Finish").green().bold()))
 }
